% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.emfrail.R
\name{autoplot}
\alias{autoplot}
\alias{autoplot.emfrail}
\title{Plots for emfrail objects using \code{ggplot2}}
\usage{
autoplot(object, ...)

\method{autoplot}{emfrail}(object, type = c("hist", "hr", "pred", "frail"),
  newdata = NULL, lp = NULL, quantity = "cumhaz", marg_cond = "both",
  conf_int = "adjusted", ...)
}
\arguments{
\item{object}{\code{emfrail} object, typically result of \code{emfrail()}}

\item{...}{Further arguments to be passed on to `ggplot`}

\item{type}{One (or more) of \code{hist} for a histogram of the estimated frailty values,
\code{hr} for a plot of the conditional and marginal hazard ratio between two cases,
\code{pred} for the predicted conditional and marginal cumulative hazard or survival for one case,
\code{frail} for a plot of the ordered frailty estimates with confidence intervals, where available.}

\item{newdata}{A \code{data.frame} with values of the covariates. For \code{type == "hr"} the hazard ratio
between the first two rows of \code{newdata} is calculated. For \code{type == "pred"} the prediction
for the first row of \code{newdata} is calculated.}

\item{lp}{A numeric vector of values of the linear predictor, each corresponding to a case. For \code{type == "hr"} the hazard ratio
between the first two values of \code{lp} is calculated. For \code{type == "pred"} the prediction
for the first value of \code{lp} is calculated.}

\item{quantity}{For \code{type == "pred"} the predicted quantity; see \code{\link{predict.emfrail}}}

\item{marg_cond}{For \code{type == "pred"} the type of predicted quantity; see \code{\link{predict.emfrail}}}

\item{conf_int}{For \code{type == "pred"} the type of confidence intervals; see \code{\link{predict.emfrail}}}
}
\value{
A list of \code{ggplot2} objects corresponding to the required plots, or one \code{ggplot2} if only one plot is selected
}
\description{
Plots for emfrail objects using \code{ggplot2}
}
\examples{
mod_rec <- emfrail(bladder1, Surv(start, stop, status) ~ treatment + number + cluster(id))
summary(mod_rec)
# cumulative hazard
# Note: this individual has number = 0, which does not exist in the data
autoplot(mod_rec, type = "pred", lp = 0)

# survival, although not very meaningful with recurrent events
autoplot(mod_rec, type = "pred", lp = 0, quantity = "survival")


# For an individual with number == 2
autoplot(mod_rec, type = "pred", newdata = data.frame(treatment = "placebo", number = 2))

# hazard ratio between an individual with 0 and with 2 recurrences at baseline
# the marginal hazard ratio is "pulled" towards 1:

autoplot(mod_rec, type = "hr",
newdata = data.frame(treatment = c("placebo", "placebo"), number = c(0, 2)))

# hazard ratio with the stable distribution:
\dontrun{
mod_rec_stab <- emfrail(bladder1,
                        Surv(start, stop, status) ~ treatment + number + cluster(id),
                        .distribution = emfrail_distribution(dist = "stable"))

autoplot(mod_rec_stab, type = "hr",
newdata = data.frame(treatment = c("placebo", "placebo"), number = c(0, 2)))

# histogram of frailty estimates
autoplot(mod_rec_stab, type = "hist")
}
}
\seealso{
\code{\link{predict.emfrail}}, \code{\link{summary.emfrail}}, \code{\link{plot.emfrail}}.
}
