% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.emfrail.R
\name{autoplot}
\alias{autoplot}
\title{Import for the autoplot generic}
\usage{
autoplot(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Further arguments}
}
\description{
Import for the autoplot generic
}
\details{
This is just used to correctly import the S3 generic \code{autoplot} from the \code{ggplot2} package in R. In general, there seems to be an issue
when writing a method using a generic function that is not defined in the base R packages. It needs to be imported and exported as well.
This generic lies in \code{ggplot2} and is doocumented there.
}
\seealso{
\code{\link{autoplot.emfrail}}
}
\keyword{internal}
