% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_test_cph.R
\name{ca_test.coxph}
\alias{ca_test.coxph}
\title{Commenges-Andersen test for heterogeneity}
\usage{
ca_test.coxph(object)
}
\arguments{
\item{object}{A \code{coxph} object with a \code{cluster()} statement in the right-hand side of the formula.}
}
\value{
A named vector containing the test statistic, variance, and p-value
}
\description{
Commenges-Andersen test for heterogeneity
}
