% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfrail_arguments.R
\name{emfrail_control}
\alias{emfrail_control}
\title{Control parameters for emfrail}
\usage{
emfrail_control(eps = 0.001, maxit = Inf, opt_fit = TRUE,
  verbose = FALSE, fast_fit = TRUE, zerotol = 1e-04,
  opt_control = list(method = "bobyqa", itnmax = NULL, control = list()))
}
\arguments{
\item{eps}{Convergence criterion for the inner loops (the EM algorithm) for fixed frailty parameters}

\item{maxit}{Maximum number of iterations in the inner loops (the EM algorithm)}

\item{opt_fit}{Logical. Whether the outer optimization should be carried out. If \code{FALSE}, then the frailty parameter is treated as fixed.}

\item{verbose}{Logical. Whether to print out information about what is going on during maximization.}

\item{fast_fit}{Logical. Whether to go the faster route in the E step and calculate directly (when possible)}

\item{zerotol}{Lower limit for 1/frailtypar (variance in case of gamma / pvf). Below this value, the variance is taken to be 0 and the
EM is not actually performed. The log-likelihood returned by em_fit is that for the Cox model.}

\item{opt_control}{A list with arguments to be sent to the maximizer. Currently ignored, will be useful in the future.}
}
\value{
A list of arguments suitable for \code{emfrail()}!
}
\description{
Control parameters for emfrail
}
\examples{
emfrail_control()
emfrail_control(eps = 10e-7)

# A data set with very small heterogeneity
set.seed(10)
x <- sample(c(0,1/2), 500, TRUE)
tstart <- rep(0, 500)
tstop <- rexp(1 * exp(x))
status <- rep(1, 500)
id <- rep(1:100, each = 5)
dat <- data.frame(id, tstart, tstop, status, x)

# What coxph does:
library(survival)
m_cph <- coxph(Surv(tstart, tstop, status) ~ x + frailty(id), dat, ties = "breslow")
m_cph$history
# For the frailty variance, the program tries: 0, 1, 0.5, 0.005, 0.00005, etc. Stops at 5e-7.

m_ft <- emfrail(dat, Surv(tstart, tstop, status) ~ x + cluster(id))
m_ft

# The algorithm gives as frailty parameter 10587.88 which means frailty variance 1/10587.88 = 9.44e-05
# That is because by default, zerotol = 1e-04, which is the point where the algorithm decides that the frailty is 0.

# If you want the exact value of the estimate, increase the precision so the point stays somewhat interior to the parameter space.
m_ft_08 <- emfrail(dat, Surv(tstart, tstop, status) ~ x + cluster(id),
                   .control = emfrail_control(zerotol = 1e-8))
# This gives a more precise estimate, 5845410 so frailty variance 1/5845410 = 1.71e-07

}
\seealso{
\code{\link{emfrail}, \link{emfrail_distributon}}
}

