% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfrail_pll.R
\name{emfrail_pll}
\alias{emfrail_pll}
\title{Profile log-likelihood}
\usage{
emfrail_pll(.data, .formula, .distribution = emfrail_distribution(), .values)
}
\arguments{
\item{.data}{Same as in \code{emfrail}}

\item{.formula}{Same as in \code{emfrail}}

\item{.distribution}{Same as in \code{emfrail}}

\item{.values}{A vector of values on where to calculate the profile likelihood. These values are the values of "theta"}
}
\value{
The maximum log-likelihood at the specific value of theta
}
\description{
Profile log-likelihood
}

