% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estep}
\alias{Estep}
\title{Perform the E step calculations}
\usage{
Estep(c, delta, alpha, bbeta, pvfm, dist)
}
\description{
This is an inner wrapper for the C++ functions which perform the E step and is not made to be used by anybody ever.
}

