% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_transform.R
\name{laplace_transform}
\alias{laplace_transform}
\title{Laplace transform}
\usage{
laplace_transform(x, .distribution)
}
\arguments{
\item{x}{A vector of positive values where to calculate the Laplace transform}

\item{.distribution}{An \code{emfrail_distribution} object. See \code{?emfrail_distribution}.}
}
\value{
Returns a vector of the same length as \code{x} with the Laplace transform.
}
\description{
Laplace transform
}
\details{
This is a simple function which calculates the Laplace transform for the gamma, positive stable and PVF distribution.
It is intended to be used to calculate marginal quantities from an \code{emfrail} object.
}

