% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.emfrail.R
\name{plot.emfrail}
\alias{plot.emfrail}
\title{Plots for emfrail objects}
\usage{
\method{plot}{emfrail}(x, type = c("hist", "hr", "pred"), newdata = NULL,
  lp = NULL, quantity = "cumhaz", marg_cond = "both",
  conf_int = "adjusted", ...)
}
\arguments{
\item{x}{\code{emfrail} object, typically result of \code{emfrail()}}

\item{type}{One (or more) of \code{hist} for a histogram of the estimated frailty values,
\code{hr} for a plot of the conditional and marginal hazard ratio between two cases and
\code{pred} for the predicted conditional and marginal cumulative hazard or survival for one case}

\item{newdata}{A \code{data.frame} with values of the covariates. For \code{type == "hr"} the hazard ratio
between the first two rows of \code{newdata} is calculated. For \code{type == "pred"} the prediction
for the first row of \code{newdata} is calculated.}

\item{lp}{A numeric vector of values of the linear predictor, each corresponding to a case. For \code{type == "hr"} the hazard ratio
between the first two values of \code{lp} is calculated. For \code{type == "pred"} the prediction
for the first value of \code{lp} is calculated.}

\item{quantity}{For \code{type == "pred"} the predicted quantity; see \code{\link{predict.emfrail}}}

\item{marg_cond}{For \code{type == "pred"} the type of predicted quantity; see \code{\link{predict.emfrail}}}

\item{conf_int}{For \code{type == "pred"} the type of confidence intervals; see \code{\link{predict.emfrail}}}

\item{...}{Further arguments to be passed to the plot function}
}
\value{
Nothing
}
\description{
Plots for emfrail objects
}
\examples{
mod_rec <- emfrail(Surv(start, stop, status) ~ treatment + number + cluster(id), bladder1)
summary(mod_rec)

# cumulative hazard
# Note: this individual has number = 0, which does not exist in the data
plot(mod_rec, type = "pred", lp = 0)

# survival, although not very meaningful with recurrent events
plot(mod_rec, type = "pred", lp = 0, quantity = "survival")


# For an individual with number == 2
plot(mod_rec, type = "pred",
newdata = data.frame(treatment = "placebo", number = 2))

# hazard ratio between an individual with 0 and with 2 recurrences at baseline
# the marginal hazard ratio is "pulled" towards 1:

plot(mod_rec, type = "hr",
newdata = data.frame(treatment = c("placebo", "placebo"), number = c(0, 2)))

# hazard ratio with the stable distribution:
\dontrun{
mod_rec_stab <- emfrail(Surv(start, stop, status) ~ treatment + number + cluster(id),
                        bladder1,
                        distribution = emfrail_distribution(dist = "stable"))

plot(mod_rec_stab, type = "hr",
newdata = data.frame(treatment = c("placebo", "placebo"), number = c(0, 2)))

# histogram of frailty estimates
plot(mod_rec_stab, type = "hist")
}
}
\seealso{
\code{\link{predict.emfrail}}, \code{\link{summary.emfrail}}, \code{\link{autoplot.emfrail}}.
}
