% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.emfrail.R
\name{predict.emfrail}
\alias{predict.emfrail}
\title{Predictions from emfrail objects}
\usage{
\method{predict}{emfrail}(fit, lp = c(0), quantity = c("cumhaz",
  "survival"), type = c("conditional", "marginal"), conf_int = c("regular",
  "adjusted"))
}
\arguments{
\item{fit}{An \code{emfrail} fit object}

\item{lp}{A vector of linear predictor values at which to calculate the curves}

\item{conf_int}{}
}
\description{
Predictions from emfrail objects
}

