% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.emfrail.R
\name{print.emfrail}
\alias{print.emfrail}
\title{Print method for emfrail objects}
\usage{
\method{print}{emfrail}(obj)
}
\arguments{
\item{obj}{An emfrail object}
}
\value{
Nothing
}
\description{
This function is included in case someone wants to access this output quickly,
and can look into the code.
}

