% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.emfrail.R
\name{summary.emfrail}
\alias{summary.emfrail}
\title{Summary for \code{emfrail} objects}
\usage{
\method{summary}{emfrail}(object, ...)
}
\arguments{
\item{object}{An object of class \code{emfrail}}

\item{...}{Ignored}
}
\value{
An object of class \code{emfrail_summary},
which is a more human-readable list of results from an \code{emfrail} object
}
\description{
Summary for \code{emfrail} objects
}
\examples{
data("bladder")
coxph(Surv(start, stop, status) ~ treatment + frailty(id), data = bladder1, method = "breslow")
mod_gamma <- emfrail(bladder1, Surv(start, stop, status) ~ treatment + cluster(id))
summary(mod_gamma)

# plot the Empirical Bayes estimates of the frailty
data("bladder")
mod_gamma <- emfrail(bladder1, Surv(start, stop, status) ~ treatment + cluster(id))
sum_mod <- summary(mod_gamma)
library(dplyr)
library(ggplot2)

# Create a plot just with the points
pl1 <- sum_mod$z \%>\%
  arrange(z) \%>\%
  mutate(x = 1:n()) \%>\%
  ggplot(aes(x = x, y = z)) +
  geom_point()

# If the quantiles of the posterior distribution is
# known, then error bars can be added:
if(!is.null(sum_mod$z$lower_q))
  pl1 <- pl1 + geom_errorbar(aes(ymin = lower_q, ymax = upper_q), alpha = 0.5)

pl1

# The plot can be made interactive!
# ggplot2 gives a warning about the "id" aesthetic, just ignore it
pl2 <- sum_mod$z \%>\%
  arrange(z) \%>\%
  mutate(x = 1:n()) \%>\%
  ggplot(aes(x = x, y = z)) +
  geom_point(aes(id = id))

if(!is.null(sum_mod$z$lower_q))
  pl2 <- pl2 + geom_errorbar(aes(ymin = lower_q, ymax = upper_q, id = id), alpha = 0.5)

library(plotly)
ggplotly(pl2)
}
