% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfrail_aux.R
\name{dist_to_pars}
\alias{dist_to_pars}
\title{dist_to_pars}
\usage{
dist_to_pars(dist, logfrailtypar, pvfm)
}
\arguments{
\item{dist}{One of gamma, stable, pvf}

\item{logfrailtypar}{The log of theta}

\item{pvfm}{The pvfm}
}
\value{
A list with 3 elements: alpha, beta (the parameters of the Laplace transform) and dist_id.
}
\description{
dist_to_pars
}
\keyword{internal}
