% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfrail_arguments.R
\name{emfrail_control}
\alias{emfrail_control}
\title{Control parameters for emfrail}
\usage{
emfrail_control(eps = 1e-04, maxit = Inf, opt_fit = TRUE,
  verbose = FALSE, fast_fit = TRUE, se_fit = TRUE, se_adj = TRUE,
  ca_test = TRUE, only_ca_test = FALSE, lik_ci = TRUE,
  opt_control = list(interval = c(-7, 20), interval_stable = c(0, 20)))
}
\arguments{
\item{eps}{Convergence criterion for the inner loops (the EM algorithm) for fixed frailty parameters}

\item{maxit}{Maximum number of iterations in the inner loops (the EM algorithm)}

\item{opt_fit}{Logical. Whether the outer optimization should be carried out.
If \code{FALSE}, then the frailty parameter is treated as fixed and the \code{emfrail} function returns only log-likelihood. See details.}

\item{verbose}{Logical. Whether to print out information about what is going on during maximization.}

\item{fast_fit}{Logical. Whether to try to calculate the E step directly, when possible. See details.}

\item{se_fit}{Logical. Whether to calculate the variance / covariance matrix.}

\item{se_adj}{Logical. Whether to calculate the adjusted variance / covariance matrix (needs \code{se_fit == TRUE})}

\item{ca_test}{Logical. Should the Commenges-Andersen test be calculated?}

\item{only_ca_test}{Logical. Should ONLY the Commenges-Andersen test be calculated?}

\item{lik_ci}{Logical. Should likelihood-based confidence interval be calculated for the frailty parameter?}

\item{opt_control}{A list with arguments to be sent to the maximizer.
As of 0.5.5, that is optimize(), so the list should only contain a length 2 vector \code{interval}. Another field
has been added in 0.5.6, \code{interval_stable} that overrides \code{interval} for the positive stable distribution;
the reason behind this is that the algorithm tends to choke for small values, even if the maximum likelihood is
really far away.}
}
\value{
An object of the type \code{emfrail_control}.
}
\description{
Control parameters for emfrail
}
\details{
The \code{fast_fit} option make a difference when the distribution is gamma (with or without left truncation) or
inverse Gaussian, i.e. pvf with m = -1/2 (without left truncation). For all the other scenarios, the fast_fit option will
automatically be changed to FALSE. When the number of events in a cluster / individual is not very small, the cases for which
fast fitting is available will show an improvement in performance.
}
\examples{
emfrail_control()
emfrail_control(eps = 10e-7)

}
\seealso{
\code{\link{emfrail}}, \code{\link{emfrail_distribution}}, \code{\link{emfrail_pll}}
}
