% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfrail_arguments.R
\name{emfrail_distribution}
\alias{emfrail_distribution}
\title{Distribution parameters for emfrail}
\usage{
emfrail_distribution(dist = "gamma", frailtypar = 2, pvfm = -1/2)
}
\arguments{
\item{dist}{One of 'gamma', 'stable' or 'pvf'.}

\item{frailtypar}{A starting value for the 'outer' maximization with respect to the frailty parameter \eqn{\theta}. Must be positive.}

\item{pvfm}{Only relevant if \code{dist = 'pvf'} is used. It determines which PVF distribution should be used. Must be  larger than -1 and not equal to 0.}
}
\value{
A list of arguments suitable for \code{emfrail()}!
}
\description{
Distribution parameters for emfrail
}
\note{
The \code{frailtypar} argument must be positive.
}
\examples{
emfrail_distribution()
emfrail_distribution(dist = 'pvf', frailtypar = 1.5, pvfm = 0.5)
}
\seealso{
\code{\link{emfrail}, \link{emfrail_control}}
}

