% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfrail_aux.R
\name{laplace_transform}
\alias{laplace_transform}
\title{Laplace transform calculation}
\usage{
laplace_transform(x, distribution)
}
\arguments{
\item{x}{A vector of positive values where to calculate the Laplace transform}

\item{distribution}{An \code{emfrail_dist} object. See \code{?emfrail_dist}.}
}
\value{
A vector of the same length as \code{x} with the Laplace transform of \code{x}
}
\description{
Laplace transform calculation
}
\details{
This is a simple function which calculates the Laplace transform for the gamma, positive stable and PVF distribution.
It is intended to be used to calculate marginal quantities from an \code{emfrail} object.
Note that the \code{left_truncation} argument is ignored here;
the marginal survival or hazard are given for the Laplace transform of a baseline subject entered at time 0.
}
\keyword{internal}
