% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.emfrail.R
\name{predict.emfrail}
\alias{predict.emfrail}
\title{Predicted hazard and survival curves from an \code{emfrail} object}
\usage{
\method{predict}{emfrail}(fit, lp = c(0), quantity = c("cumhaz",
  "survival"), type = c("conditional", "marginal"), conf_int = c("regular",
  "adjusted"))
}
\arguments{
\item{fit}{An \code{emfrail} fit object}

\item{lp}{A vector of linear predictor values at which to calculate the curves.}

\item{quantity}{The quantity to be calculated for the values of \code{lp}}

\item{type}{The type of the quantity (conditional  / marginal)}

\item{conf_int}{The type of the confidence interval (adjusted / regular)}
}
\value{
A data frame with the column \code{time} and several other columns according to the input.
By default, for each \code{lp} it will give the following columns: \code{cumhaz}, \code{survival},
\code{cumhaz_m}, \code{survival_m} for the cumulative hazard and survival, conditional and marginal.
}
\description{
Predicted hazard and survival curves from an \code{emfrail} object
}

